Trim Yajima FUN_00434e60 table access

+++ src/Battle/btlVoice.c
@@ -7439,13 +7439,13 @@ void func_002e9450(void)
     iVar6 = iVar8;
   }
   uVar4 = FUN_002bc950(iVar7,iVar9,iVar10);
-  FUN_0027ed20(uVar4,1);
+  FUN_0027ed20((u32)uVar4,1);
   uVar4 = FUN_002baf90(uVar3,iVar12,iVar6,0,0x200);
-  FUN_0027ed20(uVar4,2);
+  FUN_0027ed20((u32)uVar4,2);
   uVar5 = FUN_002bc950(0,0,0);
   *(u8 *)uVar5 = 4;
   *(u64 *)((u8 *)uVar5 + 8) = *(u64 *)((int)uVar4 + 0x58);
-  FUN_0027ed20(uVar5,1);
+  FUN_0027ed20((u32)uVar5,1);
   FUN_002b9030(uVar3);
   return;
 }
