+++ src/Event/Fcl/fclCombineMisc.c
@@ -2473,7 +2473,7 @@ u32 FUN_003d8080(int param_1,int param_2,int param_3)
 
   u32 uVar1;
 
-  u16 uVar2;
+  u32 uVar2;
 
   u8 bVar3;
